Ajustes nos metodos de deleção de usuários

+++ src/main/java/com/college/recipes_collection/services/UserService.java
@@ -26,13 +26,11 @@ public class UserService {
 
     public void createUser(UserRequestDTO request) {
         User createdUser = new User();
-
         saveUser(createdUser, request);
     }
 
     public UserResponseDTO getUserById(Long id) {
-        User user = userRepository.findById(id)
-            .orElseThrow(() -> new RuntimeException("User not found"));
+        User user = verifyIsUserExists(id);
 
         return new UserResponseDTO(
             user.getCpf(), 
@@ -55,15 +53,26 @@ public class UserService {
             .collect(Collectors.toList());
     }
 
-    public void updateUser(Long id, UserRequestDTO request) {
-        User updatedUser = userRepository.findById(id)
-            .orElseThrow(() -> new RuntimeException("User not found"));
+    public void updateUserById(Long id, UserRequestDTO request) {
+        User updatedUser = verifyIsUserExists(id);
         saveUser(updatedUser, request);
     }
 
-    public void deleteUser(Long id) {
+    public void terminateUserById(Long id) {
+        User user = verifyIsUserExists(id);
+        user.setTerminatedAt(LocalDateTime.now());
+        userRepository.save(user);
+    }
+
+    public void deleteUserById(Long id) {
         if(userRepository.existsById(id)) {
-            userRepository.deleteById(id);
+            boolean hasRelatedEntities = userHasRelatedEntities(id);
+
+            if (!hasRelatedEntities) {
+                userRepository.deleteById(id);
+            } else {
+                throw new RuntimeException("This user has some related entitie");
+            }
         } else {
             throw new RuntimeException("User not found");
         }
@@ -86,5 +95,16 @@ public class UserService {
         }
 
         return userRepository.save(user);
+    }
+
+    private User verifyIsUserExists(Long id) {
+        return userRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("User not found"));
+    }
+
+    private boolean userHasRelatedEntities(Long id) {
+        return userRepository.hasRelatedRecipes(id) ||
+                userRepository.hasRelatedReviews(id) ||
+                userRepository.hasRelatedBooks(id);
     }
 }
